# apps/backend/static/openapi.json
{
  "openapi": "3.0.2",
  "paths": {
...
    "/api/v1/users/{id}/follow": {
      "post": {
        "summary": "Follow a user",
        "tags": [
          "User"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "operationId": "followUser",
        "responses": {
          "201": {
            "description": "201",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "createdAt": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "followerId": {
                      "type": "string"
                    },
                    "followingId": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "createdAt",
                    "followerId",
                    "followingId"
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "delete": {
        "summary": "Unfollow a user",
        "tags": [
          "User"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "operationId": "unfollowUser",
        "responses": {
          "204": {
            "description": "204"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "get": {
        "summary": "Check if the authenticated user follows this user",
        "tags": [
          "User"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "operationId": "isFollowingUser",
        "responses": {
          "204": {
            "description": "204"
          },
          "404": {
            "description": "404"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/v1/users/{id}/followers": {
      "get": {
        "summary": "Get a user's followers",
        "tags": [
          "User"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "cursorCreatedAt",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "order",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "asc",
                "desc"
              ]
            }
          }
        ],
        "operationId": "getFollowers",
        "responses": {
          "200": {
            "description": "200",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string"
                          },
                          "username": {
                            "type": "string"
                          },
                          "displayName": {
                            "type": "string"
                          },
                          "avatarKey": {
                            "type": "string"
                          },
                          "bio": {
                            "type": "string"
                          },
                          "followerCount": {
                            "type": "number"
                          },
                          "createdAt": {
                            "type": "string",
                            "format": "date-time"
                          }
                        },
                        "required": [
                          "id",
                          "username",
                          "displayName",
                          "avatarKey",
                          "bio",
                          "followerCount",
                          "createdAt"
                        ]
                      }
                    },
                    "cursorSortValue": {
                      "type": "string"
                    },
                    "cursorCreatedAt": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "hasMore": {
                      "type": "boolean"
                    }
                  },
                  "required": [
                    "data",
                    "cursorSortValue",
                    "cursorCreatedAt",
                    "hasMore"
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/users/{id}/following": {
      "get": {
        "summary": "Get who a user is following",
        "tags": [
          "User"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "cursorCreatedAt",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "order",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "asc",
                "desc"
              ]
            }
          }
        ],
        "operationId": "getFollowing",
        "responses": {
          "200": {
            "description": "200",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string"
                          },
                          "username": {
                            "type": "string"
                          },
                          "displayName": {
                            "type": "string"
                          },
                          "avatarKey": {
                            "type": "string"
                          },
                          "bio": {
                            "type": "string"
                          },
                          "followerCount": {
                            "type": "number"
                          },
                          "createdAt": {
                            "type": "string",
                            "format": "date-time"
                          }
                        },
                        "required": [
                          "id",
                          "username",
                          "displayName",
                          "avatarKey",
                          "bio",
                          "followerCount",
                          "createdAt"
                        ]
                      }
                    },
                    "cursorSortValue": {
                      "type": "string"
                    },
                    "cursorCreatedAt": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "hasMore": {
                      "type": "boolean"
                    }
                  },
                  "required": [
                    "data",
                    "cursorSortValue",
                    "cursorCreatedAt",
                    "hasMore"
                  ]
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/communities/{id}/follow": {
      "post": {
        "summary": "Follow a community without joining it",
        "tags": [
          "Community"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "operationId": "followCommunity",
        "responses": {
          "201": {
            "description": "201",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "createdAt": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "followerId": {
                      "type": "string"
                    },
                    "communityId": {
                      "type": "string",
                      "format": "uuid"
                    }
                  },
                  "required": [
                    "createdAt",
                    "followerId",
                    "communityId"
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "delete": {
        "summary": "Unfollow a community",
        "tags": [
          "Community"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "operationId": "unfollowCommunity",
        "responses": {
          "204": {
            "description": "204"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "get": {
        "summary": "Check if the authenticated user follows this community",
        "tags": [
          "Community"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "operationId": "isFollowingCommunity",
        "responses": {
          "204": {
            "description": "204"
          },
          "404": {
            "description": "404"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
  },
  "info": {
    "version": "1.0.0",
    "title": "Nexus REST API - Documentation",
    "description": "Nexus REST API - Documentation"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Local Server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "x-service-token": {
        "type": "apiKey",
        "name": "x-service-token",
        "in": "header"
      }
    }
  }
}